requirement.txt
flask
google-generativeai
azure-identity
azure-mgmt-resource
azure-mgmt-storage
python-dotenv
